pyproject.toml
[project]
name = "role-tracker-ai"
version = "0.1.0"
description = "Daily automated job-matching pipeline with tailored cover letters and email digest"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "openai>=1.0",        # Phase 3: resume embeddings (text-embedding-3-small)
    "anthropic>=0.25",    # Phase 4: cover letter tailoring agent (claude-sonnet-4-6)
    "python-dotenv>=1.0", # Load .env files for local dev
    "httpx>=0.27",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pypdf>=4.0",
    # Phase 5: web API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",  # required for file uploads
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/role_tracker"]

[tool.ruff]
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
